+++ KalturaClient/Classes/Services/ResponseProfileService.swift
@@ -69,7 +69,7 @@ public final class ResponseProfileService{
 	}
 
 	/**  Clone an existing response profile  */
-	public static func clone(id: Int, profile: ResponseProfile) -> RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, CloneTokenizer> {
+	public static func clone(id: Int64, profile: ResponseProfile) -> RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, CloneTokenizer> {
 		let request: RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, CloneTokenizer> = RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, CloneTokenizer>(service: "responseprofile", action: "clone")
 			.setBody(key: "id", value: id)
 			.setBody(key: "profile", value: profile)
@@ -87,7 +87,7 @@ public final class ResponseProfileService{
 	}
 
 	/**  Delete response profile by id  */
-	public static func delete(id: Int) -> NullRequestBuilder {
+	public static func delete(id: Int64) -> NullRequestBuilder {
 		let request: NullRequestBuilder = NullRequestBuilder(service: "responseprofile", action: "delete")
 			.setBody(key: "id", value: id)
 
@@ -104,7 +104,7 @@ public final class ResponseProfileService{
 	}
 
 	/**  Get response profile by id  */
-	public static func get(id: Int) -> RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, GetTokenizer> {
+	public static func get(id: Int64) -> RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, GetTokenizer> {
 		let request: RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, GetTokenizer> = RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, GetTokenizer>(service: "responseprofile", action: "get")
 			.setBody(key: "id", value: id)
 
@@ -176,7 +176,7 @@ public final class ResponseProfileService{
 	}
 
 	/**  Update response profile by id  */
-	public static func update(id: Int, updateResponseProfile: ResponseProfile) -> RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, UpdateTokenizer> {
+	public static func update(id: Int64, updateResponseProfile: ResponseProfile) -> RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, UpdateTokenizer> {
 		let request: RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, UpdateTokenizer> = RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, UpdateTokenizer>(service: "responseprofile", action: "update")
 			.setBody(key: "id", value: id)
 			.setBody(key: "updateResponseProfile", value: updateResponseProfile)
@@ -200,7 +200,7 @@ public final class ResponseProfileService{
 	}
 
 	/**  Update response profile status by id  */
-	public static func updateStatus(id: Int, status: ResponseProfileStatus) -> RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, UpdateStatusTokenizer> {
+	public static func updateStatus(id: Int64, status: ResponseProfileStatus) -> RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, UpdateStatusTokenizer> {
 		let request: RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, UpdateStatusTokenizer> = RequestBuilder<ResponseProfile, ResponseProfile.ResponseProfileTokenizer, UpdateStatusTokenizer>(service: "responseprofile", action: "updateStatus")
 			.setBody(key: "id", value: id)
 			.setBody(key: "status", value: status.rawValue)
